Cleanup Backup Utility

+++ src/ChurchCRM/FileSystemUtils.php
@@ -4,17 +4,20 @@ namespace ChurchCRM
 {
   class FileSystemUtils
   {
-      public static function recursiveRemoveDirectory($directory)
+      public static function recursiveRemoveDirectory($directory,$retainParentFolderAndFiles=false)
       {
           //sourced from http://stackoverflow.com/questions/11267086/php-unlink-all-files-within-a-directory-and-then-deleting-that-directory
         foreach (glob("{$directory}/*") as $file) {
             if (is_dir($file)) {
-                self::recursiveRemoveDirectory($file);
-            } else {
+                self::recursiveRemoveDirectory($file,false);
+            } else if (!$retainParentFolderAndFiles) {
                 unlink($file);
             }
         }
+        if (!$retainParentFolderAndFiles) {
           rmdir($directory);
+          
+        }
       }
 
       public static function recursiveCopyDirectory($src, $dst)
